+++ drift/lib/src/ffi/database.dart
@@ -182,7 +182,8 @@ class _VmDelegate extends DatabaseDelegate {
   @override
   Future<void> runBatched(BatchedStatements statements) async {
     final prepared = [
-      for (final stmt in statements.statements) _db.prepare(stmt),
+      for (final stmt in statements.statements)
+        _db.prepare(stmt, checkNoTail: true),
     ];
 
     for (final application in statements.arguments) {
@@ -202,7 +203,7 @@ class _VmDelegate extends DatabaseDelegate {
     if (args.isEmpty) {
       _db.execute(statement);
     } else {
-      final stmt = _db.prepare(statement);
+      final stmt = _db.prepare(statement, checkNoTail: true);
       stmt.execute(args);
       stmt.dispose();
     }
@@ -227,7 +228,7 @@ class _VmDelegate extends DatabaseDelegate {
 
   @override
   Future<QueryResult> runSelect(String statement, List<Object?> args) async {
-    final stmt = _db.prepare(statement);
+    final stmt = _db.prepare(statement, checkNoTail: true);
     final result = stmt.select(args);
     stmt.dispose();